vm: fix some size_t/uint64_t issues (6+)

+++ src/cshared/stream.h
@@ -135,7 +135,7 @@ struct Stream
     {
         return buffer.data;
     }
-    size_t raw_base_length() const
+    uint64_t raw_base_length() const
     {
         return buffer.length;
     }
